Split record repository and controllers
