+++ Media.Application/Contracts/Repositories/IAlbumsRepository.cs
@@ -1,0 +1,19 @@
+using Media.Application.Models;
+using System;
+using System.Collections.Generic;
+
+namespace Media.Application.Contracts.Repositories;
+
+public interface IAlbumsRepository
+{
+  /// <summary>
+  /// Loads list of albums.
+  /// </summary>
+  /// <param name="filter">Albums will be filtered by given filter</param>
+  /// <param name="filterByGroups">True if records will be filtered by groups.</param>
+  /// <param name="clientGroups">List of groups for which the records will be loaded.</param>
+  /// <param name="skip">Elements to be skipped. default <see cref="List.Skip"/></param>
+  /// <param name="take">Elements to be loaded in one chunk. Default <see cref="List.Take"/></param>
+  /// <returns><see cref="Albums"/></returns>
+  Albums ListAlbums(string? filter, bool filterByGroups, IEnumerable<Guid> clientGroups, int skip = Constants.List.Skip, int take = Constants.List.Take);
+}
